requirements.txt
Flask==2.1.3
requests==2.28.2
python-dotenv==1.0.0
openai>=1.0.0
twilio==8.0.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
redis==4.5.1
supabase==1.0.0
numpy==1.26.4
langdetect==1.0.9
Werkzeug==2.3.8
